feat: create reusable search bar component
